Refactor cli serve command
