--- packages/cli/src/commands/serveHandler.js
@@ -1,123 +1,0 @@
-import chalk from 'chalk'
-
-import {
-  coerceRootPath,
-  createFastifyInstance,
-  redwoodFastifyAPI,
-  redwoodFastifyWeb,
-} from '@redwoodjs/fastify'
-import { getConfig } from '@redwoodjs/project-config'
-
-export const apiServerHandler = async (options) => {
-  const { port, socket, apiRootPath } = options
-  const tsApiServer = Date.now()
-
-  console.log(chalk.dim.italic('Starting API Server...'))
-
-  const fastify = createFastifyInstance()
-
-  process.on('exit', () => {
-    fastify?.close()
-  })
-
-  await fastify.register(redwoodFastifyAPI, {
-    redwood: {
-      ...options,
-    },
-  })
-
-  let listenOptions
-
-  if (socket) {
-    listenOptions = { path: socket }
-  } else {
-    listenOptions = {
-      port,
-      host: process.env.NODE_ENV === 'production' ? '0.0.0.0' : '::',
-    }
-  }
-
-  fastify.listen(listenOptions)
-
-  fastify.ready(() => {
-    fastify.log.trace(
-      { custom: { ...fastify.initialConfig } },
-      'Fastify server configuration'
-    )
-    fastify.log.trace(`Registered plugins \n${fastify.printPlugins()}`)
-    console.log(chalk.italic.dim('Took ' + (Date.now() - tsApiServer) + ' ms'))
-
-    const on = socket
-      ? socket
-      : chalk.magenta(`http://localhost:${port}${apiRootPath}`)
-
-    console.log(`API listening on ${on}`)
-    const graphqlEnd = chalk.magenta(`${apiRootPath}graphql`)
-    console.log(`GraphQL endpoint at ${graphqlEnd}`)
-
-    sendProcessReady()
-  })
-}
-
-export const bothServerHandler = async (options) => {
-  const { port, socket } = options
-  const tsServer = Date.now()
-
-  console.log(chalk.italic.dim('Starting API and Web Servers...'))
-
-  const fastify = createFastifyInstance()
-
-  process.on('exit', () => {
-    fastify?.close()
-  })
-
-  await fastify.register(redwoodFastifyWeb, {
-    redwood: {
-      ...options,
-    },
-  })
-
-  const apiRootPath = coerceRootPath(getConfig().web.apiUrl)
-
-  await fastify.register(redwoodFastifyAPI, {
-    redwood: {
-      ...options,
-      apiRootPath,
-    },
-  })
-
-  let listenOptions
-
-  if (socket) {
-    listenOptions = { path: socket }
-  } else {
-    listenOptions = {
-      port,
-      host: process.env.NODE_ENV === 'production' ? '0.0.0.0' : '::',
-    }
-  }
-
-  fastify.listen(listenOptions)
-
-  fastify.ready(() => {
-    console.log(chalk.italic.dim('Took ' + (Date.now() - tsServer) + ' ms'))
-
-    const on = socket
-      ? socket
-      : chalk.magenta(`http://localhost:${port}${apiRootPath}`)
-
-    const webServer = chalk.green(`http://localhost:${port}`)
-    const apiServer = chalk.magenta(`http://localhost:${port}`)
-    console.log(`Web server started on ${webServer}`)
-    console.log(`API serving from ${apiServer}`)
-    console.log(`API listening on ${on}`)
-    const graphqlEnd = chalk.magenta(`${apiRootPath}graphql`)
-    console.log(`GraphQL endpoint at ${graphqlEnd}`)
-
-    sendProcessReady()
-  })
-}
-
-function sendProcessReady() {
-  return process.send && process.send('ready')
-}
